pyproject.toml
[project]
name = "taxoadapt_only_gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.10.5",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "async-timeout==4.0.3",
    "attrs==25.4.0",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cloudpickle==3.1.2",
    "compressed-tensors==0.6.0",
    "datasets==4.4.2",
    "dill==0.4.0",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "einops==0.8.1",
    "et-xmlfile==2.0.0",
    "exceptiongroup==1.3.1",
    "fastapi==0.128.0",
    "filelock==3.20.2",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "gguf==0.10.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "huggingface-hub==0.36.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "inquirerpy==0.3.4",
    "interegular==0.3.3",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "joblib==1.4.2",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "lark==1.3.1",
    "llvmlite==0.46.0",
    "lm-format-enforcer==0.10.6",
    "markupsafe==3.0.3",
    "mistral-common==1.8.8",
    "mpmath==1.3.0",
    "msgpack==1.1.2",
    "msgspec==0.20.0",
    "multidict==6.7.0",
    "multiprocess==0.70.18",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "numba==0.63.1",
    "numpy==1.25.0",
    "nvidia-cublas-cu12==12.1.3.1",
    "nvidia-cuda-cupti-cu12==12.1.105",
    "nvidia-cuda-nvrtc-cu12==12.1.105",
    "nvidia-cuda-runtime-cu12==12.1.105",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.0.2.54",
    "nvidia-curand-cu12==10.3.2.106",
    "nvidia-cusolver-cu12==11.4.5.107",
    "nvidia-cusparse-cu12==12.1.0.106",
    "nvidia-ml-py==13.590.44",
    "nvidia-nccl-cu12==2.20.5",
    "nvidia-nvjitlink-cu12==12.9.86",
    "nvidia-nvtx-cu12==12.1.105",
    "openai==2.14.0",
    "opencv-python-headless==4.11.0.86",
    "openpyxl==3.1.5",
    "outlines==0.0.46",
    "packaging==25.0",
    "pandas==2.3.3",
    "partial-json-parser==0.2.1.1.post7",
    "pfzy==0.3.4",
    "pillow==12.1.0",
    "prometheus-client==0.23.1",
    "prometheus-fastapi-instrumentator==7.1.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==6.33.2",
    "psutil==7.2.1",
    "py-cpuinfo==9.0.0",
    "pyairports",
    "pyarrow==22.0.0",
    "pycountry==24.6.1",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-extra-types==2.11.0",
    "pypdf==5.8.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "ray==2.53.0",
    "referencing==0.37.0",
    "regex==2025.11.3",
    "requests==2.32.5",
    "rpds-py==0.30.0",
    "safetensors==0.7.0",
    "scikit-learn==1.3.2",
    "scipy==1.15.3",
    "sentence-transformers==3.0.1",
    "sentencepiece==0.2.1",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.50.0",
    "streamlit>=1.53.1",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "tiktoken==0.7.0",
    "tokenizers==0.20.3",
    "torch==2.4.0",
    "torchvision==0.19.0",
    "tqdm==4.66.5",
    "transformers==4.46.3",
    "triton==3.0.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "unidecode==1.3.8",
    "urllib3==2.6.2",
    "uvicorn==0.40.0",
    "uvloop==0.22.1",
    "vllm==0.6.3.post1",
    "watchfiles==1.1.1",
    "wcwidth==0.2.14",
    "websockets==15.0.1",
    "xformers==0.0.27.post2",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
]

[tool.uv.workspace]
members = [
    "pyairports_fix",
]

[tool.uv.sources]
pyairports = { workspace = true }
